streamlit/requirements.txt
altair 
pandas
streamlit
pymongo
plotly-express
scipy